requirements.txt
django
djangorestframework
django-cors-headers
drf_yasg
djangorestframework-simplejwt
djoser